More warning fixes...
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7926 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/User/DictionaryReader/AppController.m
@@ -97,7 +97,7 @@ NSDictionary* normalAttributes;
 	// YES if and only if we are inside a link
 	BOOL inLink = NO;
   
-	unsigned nextBracketIdx;
+	NSInteger nextBracketIdx;
   
 	while (index < strLength) 
 	{
